Better way to check connected monitors.
This way you check how much monitors are connected to your computer, and choose what you want in any case.

Variables:
monitors_status: list of connections, it just shows "connected" or "disconnected" status. (This can be improved by making a dictionary with the connection name, like 'HDMI-A': 'connected', if you want to know the name of each device).